+++ src/amg/HISTORY.rst
@@ -52,3 +52,7 @@ Release History
 1.3.1
 ++++++
 * `az grafana delete`: bump azure-mgmt-authorization package version
+
+1.3.2
+++++++
+* Revert to vendored SDK to fix an issue caused by Homebrew dependencies for Mac users